+++ packages/ckb-sdk-utils/src/index.ts
@@ -1,7 +1,12 @@
 import * as util from 'util'
 import crypto from './crypto'
+import { serializeScript } from './serialization/script'
+import { serializeRawTransaction } from './serialization/transaction'
 
 export * from './address'
+export * from './serialization'
+export { serializeScript } from './serialization/script'
+export { serializeRawTransaction } from './serialization/transaction'
 
 declare const TextDecoder: any // will be removed when Node@11 becomes LTS
 declare const TextEncoder: any // will be removed when Node@11 becomes LTS
@@ -20,6 +25,7 @@ export const hexToBytes = (rawhex: any) => {
   for (let c = 0; c < hex.length; c += 2) {
     bytes.push(parseInt(hex.substr(c, 2), 16))
   }
+
   return new Uint8Array(bytes)
 }
 
@@ -42,22 +48,38 @@ export const utf8ToBytes = (str: string) => textEncoder.encode(str)
 
 export const utf8ToHex = (str: string) => bytesToHex(utf8ToBytes(str))
 
-export const scriptToHash = ({ codeHash = '', args = [], hashType = 'data' }: CKBComponents.Script) => {
+export const scriptToHash = (script: CKBComponents.Script) => {
+  if (!script) throw new Error('Script is required')
+  const serializedScript = serializeScript(script)
   const s = blake2b(32, null, null, PERSONAL)
-  if (codeHash) {
-    s.update(hexToBytes(codeHash.replace(/^0x/, '')))
-  }
-
-  if (hashType === 'data') {
-    s.update(Buffer.from([0x0]))
-  } else {
-    s.update(Buffer.from([0x1]))
-  }
-
-  if (args && args.length) {
-    args.forEach(arg => (typeof arg === 'string' ? s.update(hexToBytes(arg)) : s.update(arg)))
-  }
-
+  s.update(hexToBytes(serializedScript))
   const digest = s.digest('hex')
   return digest as string
+}
+
+export const rawTransactionToHash = (rawTransaction: CKBComponents.RawTransaction) => {
+  if (!rawTransaction) throw new Error('Raw transaction is required')
+  const serializedRawTransaction = serializeRawTransaction(rawTransaction)
+  const s = blake2b(32, null, null, PERSONAL)
+  s.update(hexToBytes(serializedRawTransaction))
+  const digest = s.digest('hex')
+  return digest as string
+}
+
+const reverseString = (str: string) =>
+  str
+    .split('')
+    .reverse()
+    .join('')
+export const toHexInLittleEndian = (int: number | string, paddingBytes: number = 4) => {
+  if (Number.isNaN(+int)) {
+    throw new TypeError('The input cannot be converted to a number')
+  }
+  const hex = (+int).toString(16)
+  const reversedHex = reverseString(hex)
+  const frags = reversedHex.match(/\w{1,2}/g) || []
+  return frags
+    .map(frag => reverseString(frag.padEnd(2, '0')))
+    .join('')
+    .padEnd(paddingBytes * 2, '0')
 }